ci: correctly read and apply aws env for minikube in Jenkins

+++ deploy/dev/k8s/utils/parse_cld.sh
@@ -15,31 +15,53 @@ if [[ "${AIS_BACKEND_PROVIDERS}" == *aws* ]]; then
     fi
     # to get proper tilde expansion
     aws_env="${aws_env/#\~/$HOME}"
-    temp_file="$aws_env/credentials"
-    if [ -f $"$temp_file" ]; then
-        cp $"$temp_file"  ${LOCAL_AWS}
-    else
-        echo "No AWS credentials file found in specified directory. Exiting..."
-        exit 1
-    fi
-
-    # By default, the region field is found in the aws config file.
-    # Sometimes it is found in the credentials file.
-    if [ $(cat "$temp_file" | grep -c "region") -eq 0 ]; then
-        temp_file="$aws_env/config"
-        if [ -f $"$temp_file" ] && [ $(cat $"$temp_file" | grep -c "region") -gt 0 ]; then
-            grep region "$temp_file" >> ${LOCAL_AWS}
+    if [[ -d ${aws_env} ]]; then # Directory that contains `credentials` and `config` files.
+        temp_file="$aws_env/credentials"
+        if [ -f $"$temp_file" ]; then
+            cp $"$temp_file"  ${LOCAL_AWS}
         else
-            echo "No region config field found in aws directory. Exiting..."
+            echo "No AWS credentials file found in specified directory. Exiting..."
             exit 1
         fi
+
+        # By default, the region field is found in the aws config file.
+        # Sometimes it is found in the credentials file.
+        if [ $(cat "$temp_file" | grep -c "region") -eq 0 ]; then
+            temp_file="$aws_env/config"
+            if [ -f $"$temp_file" ] && [ $(cat $"$temp_file" | grep -c "region") -gt 0 ]; then
+                grep region "$temp_file" >> ${LOCAL_AWS}
+            else
+                echo "No region config field found in aws directory. Exiting..."
+                exit 1
+            fi
+        fi
+
+        # TODO: The following should happen only for AWS. This is because
+        # in the yaml templates this field is hardcoded. Can be solved if
+        # templating engine like helm, kustomize is used.
+        if kubectl get secrets | grep aws > /dev/null 2>&1; then
+            kubectl delete secret aws-credentials
+        fi
+        kubectl create secret generic aws-credentials --from-file=$LOCAL_AWS
+    elif [[ -f $aws_env ]]; then # File contains env variables.
+        source ${aws_env}
+        cat > ${LOCAL_AWS} <<- EOM
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: aws-credentials
+  namespace: default
+data:
+  AWS_ACCESS_KEY_ID: ${AWS_ACCESS_KEY_ID}
+  AWS_SECRET_ACCESS_KEY: ${AWS_SECRET_ACCESS_KEY}
+  AWS_DEFAULT_REGION: ${AWS_DEFAULT_REGION}
+EOM
+
+        if kubectl get configmap | grep aws > /dev/null 2>&1; then
+            kubectl delete configmap aws-credentials
+        fi
+        kubectl create -f ${LOCAL_AWS}
+    else
+      echo "File nor directory exists under: ${aws_env}"
     fi
 fi
-
-# TODO: The following should happen only for AWS. This is because
-# in the yaml templates this field is hardcoded. Can be solved if
-# templating engine like helm, kustomize is used.
-if kubectl get secrets | grep aws > /dev/null 2>&1; then
-    kubectl delete secret aws-credentials
-fi
-kubectl create secret generic aws-credentials --from-file=$LOCAL_AWS
